requirements.txt
django
django-filter
djangorestframework
drf-spectacular
drf-spectacular[sidecar]
django-environ
gunicorn
pyyaml
uritemplate
